short traces?
